requirements.txt
git addcertifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.0
requests==2.28.2
urllib3==1.26.14
